WIP: Only prompt for root when necessary
It always bothers me when apps prompt for permissions when it's not a
direct result of me trying to do something that requires those
permissions.

There is an issue, however: Adjusting the sliders on the home screen
will send many updates to the filter service, resulting in a million
root prompts, if permissions are not already granted, which is a
terrible UX. So, we can't allow adjusting those sliders until we're sure
we have root.

At the moment, I think I will make the prompt happen when the
MainActivity opens, only if root mode is enabled. It's not ideal, but
it's the easiest to implement for now.